Priority replay buffer fixes

+++ rl_modules/prioritised_replay_buffer.py
@@ -9,7 +9,7 @@ class ReplayBuffer:
     def __init__(self, obs_dim: int, size: int, batch_size: int = 32):
         self.obs_buf = np.zeros([size, *obs_dim], dtype=np.float32)
         self.next_obs_buf = np.zeros([size, *obs_dim], dtype=np.float32)
-        self.acts_buf = np.zeros([size], dtype=np.float32)
+        self.acts_buf = np.zeros([size], dtype=np.int64)
         self.rews_buf = np.zeros([size], dtype=np.float32)
         self.done_buf = np.zeros(size, dtype=np.float32)
         self.max_size, self.batch_size = size, batch_size
